+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - DF-811 Add support for upsert
 ### Changed
 - DF-1028 Ignored JWT explicitly when making login requests
+- Moved database configuration to service and the new NoDbModel-based handler
 ### Fixed
 - DF-1005 Fixed resource level RBAC for swagger docs
 - DF-955 Added PUT section on api docs for system/{resource} and system/{resource}/{id}